requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary==1.41.0
bcrypt==3.2.2

redis==5.0.8
fakeredis==2.23.3
jinja2==3.1.4
sphinx==7.4.7
pytest==8.3.3
pytest-cov==5.0.0
httpcore==1.0.9
httpx==0.28.1
